requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
transformers==4.41.2
torch==2.3.1
spacy==3.7.5
groq==1.0.0
trafilatura==1.12.0
python-dotenv==1.0.1
pydantic==2.7.4
requests==2.32.3